requirements.txt
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
jedi==0.17.2
lxml==4.5.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.5
Pygments==2.6.1
python-docx==0.8.10
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.5
